+++ src/files/cp.js
