+++ src/files/ls-readable-stream.js
