+++ src/files/stat.js
